services/ui/requirements.txt
streamlit
paho-mqtt
pandas